requirements.txt
pandas 
numpy 
plotly
altair
ipykernel
matplotlib
seaborn
openpyxl
python-dotenv

# Database (PostgreSQL)
psycopg2-binary
sqlalchemy
